Updating source model
I have no idea why, but I used a source model for a liquid jet for what should be a gaseous jet? While acetylene can be a liquid in the container, a leak from a hole in an acetylene line would be a gas.

While I was at it I cleaned up some of the code formatting such that it will now run. e.g. previously I used the constant `cd` but that now conflicts with the `cd()` function. It's fun how these notebooks from even a year ago stop working.