The Ideal Gas Law - bane of my existence
Hilariously, for a post on the ideal gas law, I screwed it up right at the beginning when calculating the initial mass density in the tank. (I had the ratio of tank pressure to ambient pressure backwards). The rest of it works fine, but this explains why the blowdown was unbelievably fast.